# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hotel Operations Intelligence Agent API",
    "description": "�Ƶ�������Ӫ Agent ƽ̨ API������д������ѭ���Կɽ��͡�ִ�п�����������ɹ���ʧ�ܿɻָ�������ԭ��",
    "version": "0.3.0"
  },
  "paths": {
    "/api/v2/platform/capabilities": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "List Capabilities",
        "operationId": "list_capabilities_api_v2_platform_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "type": "object",
                  "title": "Response List Capabilities Api V2 Platform Capabilities Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/knowledge/search": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Planned Capability",
        "operationId": "planned_capability_api_v2_knowledge_search_post",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/measurements/{measurement_id}": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Planned Capability",
        "operationId": "planned_capability_api_v2_measurements__measurement_id__get",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/content/publish": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Planned Capability",
        "operationId": "planned_capability_api_v2_content_publish_post",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/content/generate": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Planned Capability",
        "operationId": "planned_capability_api_v2_content_generate_post",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/executions": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Planned Capability",
        "operationId": "planned_capability_api_v2_executions_post",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/approvals": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Planned Capability",
        "operationId": "planned_capability_api_v2_approvals_post",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/revenue/recommendations": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Planned Capability",
        "operationId": "planned_capability_api_v2_revenue_recommendations_post",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/forecast/demand": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Planned Capability",
        "operationId": "planned_capability_api_v2_forecast_demand_post",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/market/events": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Planned Capability",
        "operationId": "planned_capability_api_v2_market_events_get",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/market/rate-shop": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Planned Capability",
        "operationId": "planned_capability_api_v2_market_rate_shop_post",
        "parameters": [
          {
            "name": "measurement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measurement Id"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/dashboard/overview": {
      "get": {
        "tags": [
          "��Ӫ��ʻ��"
        ],
        "summary": "��ѯ��Ӫ��ʻ�ո���",
        "operationId": "get_overview_api_v2_dashboard_overview_get",
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "default": "hotel-001",
              "title": "Property Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/ops/tasks": {
      "post": {
        "tags": [
          "��Ӫ����"
        ],
        "summary": "������Ӫ�������������",
        "operationId": "create_task_api_v2_ops_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsTask"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "��Ӫ����"
        ],
        "summary": "��ѯ��Ӫ�����б�",
        "operationId": "list_tasks_api_v2_ops_tasks_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "property_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Property Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsTaskList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/ops/tasks/{task_id}": {
      "get": {
        "tags": [
          "��Ӫ����"
        ],
        "summary": "��ѯ����״̬�ͽ��",
        "operationId": "get_task_api_v2_ops_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsTask"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/ops/tasks/{task_id}/events": {
      "get": {
        "tags": [
          "��Ӫ����"
        ],
        "summary": "��ѯ�����¼���",
        "operationId": "get_task_events_api_v2_ops_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Task Events Api V2 Ops Tasks  Task Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/ops/tasks/{task_id}/approve": {
      "post": {
        "tags": [
          "��Ӫ����"
        ],
        "summary": "��׼���Բ�������",
        "operationId": "approve_task_api_v2_ops_tasks__task_id__approve_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsTask"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/ops/tasks/{task_id}/execute": {
      "post": {
        "tags": [
          "��Ӫ����"
        ],
        "summary": "ִ������׼���ܿض���",
        "operationId": "execute_task_api_v2_ops_tasks__task_id__execute_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsTask"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DashboardOverview": {
        "properties": {
          "property_id": {
            "type": "string",
            "title": "Property Id"
          },
          "data_cutoff": {
            "type": "string",
            "format": "date-time",
            "title": "Data Cutoff"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricCard"
            },
            "type": "array",
            "title": "Metrics"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/OperatingAlert"
            },
            "type": "array",
            "title": "Alerts"
          },
          "pending_approval_count": {
            "type": "integer",
            "title": "Pending Approval Count"
          }
        },
        "type": "object",
        "required": [
          "property_id",
          "data_cutoff",
          "metrics",
          "alerts",
          "pending_approval_count"
        ],
        "title": "DashboardOverview"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MetricCard": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "display_value": {
            "type": "string",
            "title": "Display Value"
          },
          "change_display": {
            "type": "string",
            "title": "Change Display"
          },
          "trend": {
            "type": "string",
            "pattern": "^(up|down|flat)$",
            "title": "Trend"
          }
        },
        "type": "object",
        "required": [
          "code",
          "label",
          "value",
          "display_value",
          "change_display",
          "trend"
        ],
        "title": "MetricCard"
      },
      "OperatingAlert": {
        "properties": {
          "severity": {
            "type": "string",
            "pattern": "^(critical|high|medium|low)$",
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "module": {
            "type": "string",
            "title": "Module"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          }
        },
        "type": "object",
        "required": [
          "severity",
          "title",
          "description",
          "module",
          "occurred_at"
        ],
        "title": "OperatingAlert"
      },
      "OpsTask": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState",
            "default": "PENDING"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "request": {
            "$ref": "#/components/schemas/TaskCreate"
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recommendation"
              },
              {
                "type": "null"
              }
            ]
          },
          "approval_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Hash"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "request"
        ],
        "title": "OpsTask"
      },
      "OpsTaskList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OpsTask"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "OpsTaskList"
      },
      "Recommendation": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "parameters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "type": "object",
            "title": "Parameters"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Refs"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "action",
          "parameters",
          "rationale",
          "confidence",
          "evidence_refs"
        ],
        "title": "Recommendation"
      },
      "TaskCreate": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "minLength": 1,
            "title": "Tenant Id"
          },
          "property_id": {
            "type": "string",
            "minLength": 1,
            "title": "Property Id"
          },
          "objective": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Objective"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "default": "revenue_recommendation"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority",
            "default": "MEDIUM"
          },
          "data_cutoff": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Cutoff"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "property_id",
          "objective"
        ],
        "title": "TaskCreate"
      },
      "TaskPriority": {
        "type": "string",
        "enum": [
          "CRITICAL",
          "HIGH",
          "MEDIUM",
          "LOW"
        ],
        "title": "TaskPriority"
      },
      "TaskState": {
        "type": "string",
        "enum": [
          "PENDING",
          "ANALYZING",
          "WAITING_APPROVAL",
          "APPROVED",
          "EXECUTED",
          "REJECTED"
        ],
        "title": "TaskState"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "��Ӫ��ʻ��",
      "description": "��Ӫ������ָ��͸澯��"
    },
    {
      "name": "��Ӫ����",
      "description": "����״̬��ת���������ܿ�ִ�С�"
    },
    {
      "name": "platform",
      "description": "ƽ̨������ģ����Լ��"
    }
  ]
}